:sparkles: Remove Webpack and compile javascript using Terser and Eleventy's new custom language features
