+++ package.json
@@ -19,21 +19,10 @@
   "license": "MIT",
   "dependencies": {
     "@11ty/eleventy": "^1.0.0-beta",
-    "@babel/core": "^7.13.14",
-    "@babel/plugin-transform-runtime": "^7.13.10",
-    "@babel/preset-env": "^7.13.12",
-    "autoprefixer": "^10.3.1",
-    "babel-loader": "^8.2.2",
-    "clean-css": "^5.1.2",
-    "cross-env": "^7.0.3",
-    "cssnano": "^5.0.7",
     "markdown-it": "^12.2.0",
-    "memfs": "^3.2.0",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.3.6",
-    "postcss-cli": "^8.3.1",
     "sass": "^1.42.1",
-    "webpack": "^5.28.0"
+    "terser": "^5.10.0"
   },
   "browserslist": [
     "last 2 versions"
